SynchronusShopping: Attempted a solution that would follow these steps
Alternate Cats, find the road that takes the shortest time and has some fish we need (be greedy)
Once all fish were found attempt Dijkstras Alg to find shortest path the last store.

This would yield incorrect results...My theory is that the greedy part is incorrect and perhaps making moves that are unwise.  My next attempt will to do pure Dijkstra but mess with the time on the roads (i.e. set them really high if the next store does not contain a fish we need).